+++ src/main/java/fr/heneriacore/event/AuthPostLoginEvent.java
@@ -1,0 +1,28 @@
+package fr.heneriacore.event;
+
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+
+import java.util.UUID;
+
+public class AuthPostLoginEvent extends Event {
+    private static final HandlerList HANDLERS = new HandlerList();
+    private final UUID uuid;
+
+    public AuthPostLoginEvent(UUID uuid) {
+        this.uuid = uuid;
+    }
+
+    public UUID getUuid() {
+        return uuid;
+    }
+
+    @Override
+    public HandlerList getHandlers() {
+        return HANDLERS;
+    }
+
+    public static HandlerList getHandlerList() {
+        return HANDLERS;
+    }
+}